Capture and return API response 'message' key on error
The 'message' field in Tapis API error reponses was not being propagated with the
HTTPError object, thus forcing consumers of AgavePy to write custom handler code
to extract the error. This is now done automatically using the 'message' arg for
HTTPError.

This new function can be disabled by setting TAPIS_PY_VERBOSE_ERRORS to
a falsey value.